% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffExpr.R
\name{diffExpr}
\alias{diffExpr}
\title{Select DE result by cutoff}
\usage{
diffExpr(data1, data2, x, y, cut_pvalue, cut_fc)
}
\arguments{
\item{data1}{a dataframe of expression A}

\item{data2}{a dataframe of expression B}

\item{x}{a numeric vector of fold change result from 'FC' func}

\item{y}{a numeric vector of p value result from 'pValue' func}

\item{cut_pvalue}{pvalue cutoff}

\item{cut_fc}{fold change cutoff , value is abs(log2FC)}
}
\value{
a dataframe of differential expression genes, which contains the mean expression and fold change and p-value
}
\description{
Select DE result by cutoff
}
\examples{
de = diffExpr(data1,data2,p_value,fc,0.05,0.5)
}
