% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FC.R
\name{FC}
\alias{FC}
\title{Compute the fold change between expression A and expression B of each gene}
\usage{
FC(x, y)
}
\arguments{
\item{x}{dataframe of expression A}

\item{y}{dataframe of expression B}
}
\value{
a numeric vector
}
\description{
compute the fold change
}
\examples{
fc = FC(x,y)
}
