% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pValue.R
\name{pValue}
\alias{pValue}
\title{Compute p-value using t.test from expression A to expression B}
\usage{
pValue(x, y)
}
\arguments{
\item{x}{a dataframe of expression A}

\item{y}{a dataframe of expression B}
}
\value{
a numeric vector of pvalue between genes
}
\description{
Compute p-value using t.test from expression A to expression B
}
\examples{
pvalue = pValue(x,y)
}
